+++ src/web/assets/dashboard/DashboardAsset.php
@@ -7,9 +7,9 @@
 
 namespace craft\web\assets\dashboard;
 
-use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
 use craft\web\View;
+use yii\web\AssetBundle;
 
 /**
  * Asset bundle for the Dashboard